+++ packages/jest-repl/src/cli/index.ts
@@ -5,23 +5,23 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
- * @flow
  */
-
-import path from 'path';
 
 import Runtime from 'jest-runtime';
 import yargs from 'yargs';
 import {validateCLIOptions} from 'jest-validate';
 import {deprecationEntries} from 'jest-config';
-import {version as VERSION} from '../../package.json';
+import {Config} from '@jest/types';
 import * as args from './args';
 
-const REPL_SCRIPT = path.resolve(__dirname, './repl.js');
+const {version: VERSION} = require('../../package.json');
 
-module.exports = function() {
-  const argv = yargs.usage(args.usage).options(args.options).argv;
+const REPL_SCRIPT = require.resolve('./repl.js');
 
+export = function() {
+  const argv = <Config.Argv>yargs.usage(args.usage).options(args.options).argv;
+
+  // @ts-ignore: fix this at some point
   validateCLIOptions(argv, {...args.options, deprecationEntries});
 
   argv._ = [REPL_SCRIPT];