+++ packages/jest-validate/src/index.ts
@@ -3,8 +3,6 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
- *
- * @flow
  */
 
 import {
@@ -17,7 +15,7 @@ import validate from './validate';
 import validateCLIOptions from './validateCLIOptions';
 import {multipleValidOptions} from './condition';
 
-module.exports = {
+export = {
   ValidationError,
   createDidYouMeanMessage,
   format,